Use SciMLTesting v1.2 folder-based run_tests
