+++ Project.toml
@@ -17,8 +17,9 @@ DifferentialEquations = "7"
 ForwardDiff = "0.10, 1.4"
 Optimisers = "0.2, 0.3, 0.4"
 Random = "1"
-SafeTestsets = "0.1"
+SafeTestsets = "0.1, 1"
 SciMLBase = "1, 2, 3.1"
+SciMLTesting = "1"
 Test = "1"
 Yao = "0.8, 0.9"
 Zygote = "0.6, 0.7, 0.8"
@@ -26,10 +27,10 @@ julia = "1.10"
 
 [extras]
 DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["SafeTestsets", "Test", "DifferentialEquations", "Random", "Pkg"]
+test = ["SafeTestsets", "SciMLTesting", "Test", "DifferentialEquations", "Random"]
