+++ test/runtests.jl
@@ -1,230 +1,2 @@
-using Test
-
-const GROUP = get(ENV, "GROUP", "All")
-
-if GROUP == "QA"
-    import Pkg
-    Pkg.activate(joinpath(@__DIR__, "qa"))
-    Pkg.instantiate()
-    include(joinpath(@__DIR__, "qa", "qa.jl"))
-    exit()
-end
-
-using QuantumNLDiffEq
-using Yao: dispatch, EasyBuild, put, Z, chain, Ry, parameters, zero_state, expect,
-    parameters, nparameters, Add
-using Zygote: gradient
-using DifferentialEquations
-using Optimisers: Adam
-using Random
-
-# Generate reproducible initial parameters using a local Xoshiro RNG so the
-# variational-circuit initialization is identical across Julia versions
-# (Julia's default global RNG stream changes between releases, which made
-# the loss-threshold tests below flaky when using `dispatch(..., :random)`).
-function init_var_circuit(seed::Integer)
-    circ = EasyBuild.variational_circuit(6, 5)
-    rng = Xoshiro(seed)
-    return dispatch(circ, rand(rng, nparameters(circ)))
-end
-
-M = range(0; stop = 0.9, length = 20)
-@testset "Tests for damped oscillation equations" begin
-    function f(u, p, t)
-        λ, κ = p
-        return -1 * λ * u * (κ + tan(λ * t))
-    end
-    prob = ODEProblem(f, [1.0], (0.0, 0.9), [8.0, 0.1])
-    function loss_func(a, b)
-        return (a - b)^2
-    end
-
-    @testset "Boundary Handling Tests" begin
-        config(boundary) = DQCConfig(abh = boundary, loss = loss_func)
-
-        @testset "Test for Pinned Boundary Handling" begin
-            DQC = [
-                QuantumNLDiffEq.DQCType(
-                    afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                    fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                    cost = [Add([put(6, i => Z) for i in 1:6])],
-                    var = init_var_circuit(42), N = 6
-                ),
-            ]
-            conf = config(QuantumNLDiffEq.Pinned(2.5))
-            params = [parameters(DQC[1].var)]
-            # 600 steps (up from default 300) so convergence is not on the
-            # threshold edge: with the chosen seed and 300 steps, the Pinned
-            # loss lands near 3 on some Julia versions; doubling the iterations
-            # gives the Adam optimizer a comfortable margin under the 2.0 cap.
-            QuantumNLDiffEq.train!(DQC, prob, conf, M, params; steps = 600)
-            @test QuantumNLDiffEq.loss(DQC, prob, conf, M, params) < 2.0
-        end
-
-        @testset "Test for Floating Boundary Handling" begin
-            # seed=42 lands in a local minimum (loss ~0.85) for this ChebyshevSparse
-            # + Floating setup; seed=7 with 600 Adam steps converges well below 0.5
-            DQC = [
-                QuantumNLDiffEq.DQCType(
-                    afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                    fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                    cost = [Add([put(6, i => Z) for i in 1:6])],
-                    var = init_var_circuit(7), N = 6
-                ),
-            ]
-            conf = config(QuantumNLDiffEq.Floating())
-            params = [parameters(DQC[1].var)]
-            QuantumNLDiffEq.train!(DQC, prob, conf, M, params; steps = 600)
-            @test QuantumNLDiffEq.loss(DQC, prob, conf, M, params) < 0.5
-        end
-
-        @testset "Test for Optimised Boundary Handling" begin
-            DQC = [
-                QuantumNLDiffEq.DQCType(
-                    afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                    fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                    cost = [Add([put(6, i => Z) for i in 1:6])],
-                    var = init_var_circuit(42), N = 6
-                ),
-            ]
-            # Deterministic initial boundary value for Optimized handling
-            # (avoid the global-RNG-dependent `rand()` from the previous code).
-            conf = config(QuantumNLDiffEq.Optimized(rand(Xoshiro(42))))
-            params = [parameters(DQC[1].var)]
-            QuantumNLDiffEq.train!(DQC, prob, conf, M, params)
-            @test QuantumNLDiffEq.loss(DQC, prob, conf, M, params) < 0.5
-        end
-    end
-
-    @testset "Mapping Tests" begin
-        DQ(mapping) = [
-            QuantumNLDiffEq.DQCType(
-                afm = mapping, fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                cost = [Add([put(6, i => Z) for i in 1:6])],
-                var = init_var_circuit(42), N = 6
-            ),
-        ]
-        config = DQCConfig(abh = QuantumNLDiffEq.Floating(), loss = loss_func)
-
-        @testset "Test for Product Feature Mapping" begin
-            input = DQ(QuantumNLDiffEq.Product())
-            params = [parameters(input[1].var)]
-            QuantumNLDiffEq.train!(input, prob, config, M, params)
-            @test QuantumNLDiffEq.loss(input, prob, config, M, params) < 10
-        end
-
-        @testset "Test for Chebyshev Sparse Mapping" begin
-            # Same ChebyshevSparse + Floating setup as the Floating test above
-            # gets stuck at the same plateau under seed 42; use seed 7 and 600
-            # Adam steps to comfortably reach loss < 0.5.
-            input = QuantumNLDiffEq.DQCType(
-                afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                cost = [[Add([put(6, i => Z) for i in 1:6])]],
-                var = init_var_circuit(7), N = 6
-            )
-            params = parameters(input.var)
-            QuantumNLDiffEq.train!(input, prob, config, M, params; steps = 600)
-            @test QuantumNLDiffEq.loss(input, prob, config, M, params) < 0.5
-        end
-
-        @testset "Test for Chebyshev Tower Mapping" begin
-            input = DQ(QuantumNLDiffEq.ChebyshevTower(2))
-            params = [parameters(input[1].var)]
-            QuantumNLDiffEq.train!(input, prob, config, M, params)
-            @test QuantumNLDiffEq.loss(input, prob, config, M, params) < 0.5
-        end
-    end
-
-    @testset "Regularization Tests" begin
-        @testset "Singular Encoding" begin
-            DQC = QuantumNLDiffEq.DQCType(
-                afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                cost = [[Add([put(6, i => Z) for i in 1:6])]],
-                var = init_var_circuit(42), N = 6
-            )
-            config = DQCConfig(
-                reg = QuantumNLDiffEq.RegularisationParams(
-                    [
-                        [
-                            1.0, 0.04724630684751344, -0.7340180576454999,
-                            -0.10424985483835963, 0.5322086377394599,
-                        ],
-                    ],
-                    [
-                        0.0, 0.18947368421052632, 0.37894736842105264,
-                        0.5684210526315789, 0.7578947368421053,
-                    ],
-                    0.0
-                ),
-                abh = QuantumNLDiffEq.Floating(),
-                loss = loss_func
-            )
-            params = parameters(DQC.var)
-            QuantumNLDiffEq.tr_custom!(DQC, prob, config, M, params)
-            @show loss = QuantumNLDiffEq.loss(DQC, prob, config, M, params)
-            @test loss < 0.5
-        end
-
-        @testset "Multiple Encoding" begin
-            DQC = [
-                QuantumNLDiffEq.DQCType(
-                    afm = QuantumNLDiffEq.ChebyshevSparse(2),
-                    fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                    cost = [Add([put(6, i => Z) for i in 1:6])],
-                    var = init_var_circuit(42), N = 6
-                ),
-            ]
-            config = DQCConfig(
-                reg = QuantumNLDiffEq.RegularisationParams(
-                    [
-                        [
-                            1.0, 0.04724630684751344, -0.7340180576454999,
-                            -0.10424985483835963, 0.5322086377394599,
-                        ],
-                    ],
-                    [
-                        0.0, 0.18947368421052632, 0.37894736842105264,
-                        0.5684210526315789, 0.7578947368421053,
-                    ],
-                    0.0
-                ),
-                abh = QuantumNLDiffEq.Floating(),
-                loss = loss_func
-            )
-            params = [parameters(DQC[1].var)]
-            QuantumNLDiffEq.tr_custom!(DQC, prob, config, M, params)
-            loss = QuantumNLDiffEq.loss(DQC, prob, config, M, params)
-            @test loss < 0.5
-        end
-    end
-end
-
-@testset "Encoding multifunction system" begin
-    function f(u, p, t)
-        λ1, λ2 = p
-        return [λ1 * u[2] + λ2 * u[1], -λ1 * u[1] - λ2 * u[2]]
-    end
-    prob = ODEProblem(f, [0.5, 0.0], (0.0, 0.9), [5.0, 3.0])
-    function loss_func(a, b)
-        return (a - b)^2
-    end
-    config = DQCConfig(abh = QuantumNLDiffEq.Floating(), loss = loss_func)
-    DQC = repeat(
-        [
-            QuantumNLDiffEq.DQCType(
-                afm = QuantumNLDiffEq.ChebyshevTower(2),
-                fm = chain(6, [put(i => Ry(0)) for i in 1:6]),
-                cost = [sum([put(6, i => Z) for i in 1:6])],
-                var = init_var_circuit(42), N = 6
-            ),
-        ],
-        2
-    )
-    params = [parameters(DQC[1].var), parameters(DQC[2].var)]
-    QuantumNLDiffEq.train!(
-        DQC, prob, config, M, params; optimizer = Adam(0.02), steps = 400
-    )
-    @test QuantumNLDiffEq.loss(DQC, prob, config, M, params) < 0.05
-end
+using SciMLTesting
+run_tests()
